Restructure app to not be a RubyGem
Making this a Gem was overcomplicating the structure of the service and
introducing excessive namespacing for no reason. I was feeling too much
friction to start making the other services with this structure.

The reality is this service is not going to be deployed as a RubyGem, so
why go to the trouble to make it one.